Enable Jumbo Frames on Ethernet + Baby Jumbo frames on WAN.
The makefile for bcm_enet has been modified to force enabling the jumbo
frames feature.

Additionally, the code in the bcmxtmrt xDSL driver has been modified to
set the MTU to 1508 when in bridging mode and also when terminating PPP
locally. However, only the bridging mode has been testing, YMMV when
trying to use it on the box itself.